Rename pattern detectors
