+++ modules/mining-pipeline/prefix-span-pattern-detector/src/main/kotlin/org/cafejojo/schaapi/miningpipeline/patterndetector/prefixspan/PrefixSpanPatternDetector.kt
@@ -1,5 +1,6 @@
-package org.cafejojo.schaapi.miningpipeline.patterndetector.ccspan
+package org.cafejojo.schaapi.miningpipeline.patterndetector.prefixspan
 
+import mu.KLogging
 import org.cafejojo.schaapi.miningpipeline.Pattern
 import org.cafejojo.schaapi.miningpipeline.PatternDetector
 import org.cafejojo.schaapi.models.GeneralizedNodeComparator
@@ -7,15 +8,21 @@ import org.cafejojo.schaapi.models.Node
 import org.cafejojo.schaapi.models.PathEnumerator
 
 /**
- * Finds closed sequential patterns using the CCSpan algorithm by Zhang et. al.
+ * A pattern detector implementing the prefix span algorithm.
  */
-class PatternDetector<N : Node>(
+class PrefixSpanPatternDetector<N : Node>(
     private val minimumCount: Int,
     private val maximumSequenceLength: Int,
     private val comparator: GeneralizedNodeComparator<N>
 ) : PatternDetector<N> {
+    private companion object : KLogging()
+
     override fun findPatterns(graphs: List<N>): List<Pattern<N>> {
+        logger.info { "Detecting patterns in ${graphs.size} graphs." }
+
         val sequences = graphs.flatMap { PathEnumerator(it, maximumSequenceLength).enumerate() }
-        return CCSpan(sequences, minimumCount, comparator).findFrequentPatterns()
+        logger.info { "Found ${sequences.size} sequences in ${graphs.size} graphs." }
+
+        return PrefixSpan(sequences, minimumCount, comparator).findFrequentPatterns()
     }
 }
